requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
lxml==4.5.0
numpy==1.18.1
oandapyV20==0.6.3
pandas==1.0.1
pandas-datareader==0.8.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
